#193: abstract DesktopFilePicker base shares pickFolder/pickPhotos; inline jvmPickFolder
