+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/transfer/AwtFilePicker.kt
@@ -7,7 +7,6 @@ import ru.pocketbyte.kydra.log.KydraLog
 import ru.pocketbyte.kydra.log.info
 import ru.pocketbyte.kydra.log.wrapper.withTag
 import java.awt.FileDialog
-import javax.swing.JFileChooser
 
 private val log = KydraLog.withTag(default = "AwtFilePicker")
 
@@ -22,20 +21,4 @@ internal suspend fun awtPickFiles(windowHolder: WindowHolder): List<FileSource> 
     }
     log.info { "pickFiles (awt): ${files.size} file(s) selected" }
     return files.map { JvmPathFileSource(it.toPath()) }
-}
-
-internal suspend fun jvmPickFolder(windowHolder: WindowHolder): List<FileSource> {
-    val dir = withContext(Dispatchers.Swing) {
-        val chooser = JFileChooser()
-        chooser.fileSelectionMode = JFileChooser.DIRECTORIES_ONLY
-        val result = chooser.showOpenDialog(windowHolder.window)
-        if (result == JFileChooser.APPROVE_OPTION) chooser.selectedFile else null
-    }
-    if (dir == null) {
-        log.info { "pickFolder: cancelled" }
-        return emptyList()
-    }
-    val sources = withContext(Dispatchers.IO) { JvmFolderWalker().walk(dir) }
-    log.info { "pickFolder: ${sources.size} file(s) from ${dir.name}" }
-    return sources
 }
