+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/transfer/LinuxFilePicker.kt
@@ -1,12 +1,7 @@
 package com.tubetoast.tether.transfer
 
 internal class LinuxFilePicker(
-    private val windowHolder: WindowHolder,
-) : FilePicker {
+    windowHolder: WindowHolder,
+) : DesktopFilePicker(windowHolder) {
     override suspend fun pickFiles(): List<FileSource> = awtPickFiles(windowHolder)
-
-    override suspend fun pickFolder(): List<FileSource> = jvmPickFolder(windowHolder)
-
-    override suspend fun pickPhotos(): List<FileSource> =
-        throw UnsupportedOperationException("pickPhotos is mobile-only")
 }
